Create Perfect Number Logic in C++
When the sum of the factors of a number, excluding itself is same as the number, the number is called a PERFECT number. This piece of code contains a function that tests whether the giver number is a perfect number or not and the main() function to drive the program.